# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Salary Manager",
        "description": "CRUD&Permisson+Role system.",
        "contact": {
            "email": "Hexagon14@mail.ru"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Main API Server"
        }
    ],
    "paths": {
...
        "/api/company": {
            "get": {
                "tags": [
                    "Companies"
                ],
                "summary": "Get list of companies.",
                "description": "Returns list of companies.",
                "operationId": "getCompanyList",
                "responses": {
                    "200": {
                        "description": "Returns list of companies."
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "Bearer Token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Companies"
                ],
                "summary": "Creates company.",
                "description": "Creates company.",
                "operationId": "createCompany",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "contacts": {
                                        "type": "string"
                                    },
                                    "create_time": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "type": "Heroku",
                                    "contacts": "admin@heroku.com",
                                    "create_time": "2006-11-22 12:00:00"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Returns company."
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "Bearer Token": []
                    }
                ]
            }
        },
        "/api/company/{id}": {
            "get": {
                "tags": [
                    "Companies"
                ],
                "summary": "Get one company.",
                "description": "Returns one company.",
                "operationId": "getCompany",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Company id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns company."
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "Bearer Token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Companies"
                ],
                "summary": "Deletes company.",
                "description": "Deletes company.",
                "operationId": "deleteCompany",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Company id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Company deleted."
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "Bearer Token": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Companies"
                ],
                "summary": "Updates company.",
                "description": "Updates company parameters that were given in request.",
                "operationId": "updateCompany",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Company id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns absence."
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "Bearer Token": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "Bearer Token": {
                "type": "apiKey",
                "description": "Laravel passport security.",
                "in": "header",
                "name": "Authorization",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost/auth/login",
                        "tokenUrl": "http://localhost/auth/token",
                        "refreshUrl": "http://localhostauth/refresh",
                        "scopes": []
                    }
                }
            }
        }
    },
    "security": [
        []
    ]
}